requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.2
dill==0.3.8
Django==4.2.13
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.14.0
Flask==3.0.3
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
huggingface-hub==0.23.3
idna==3.7
image==1.5.33
importlib_metadata==8.5.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pipdeptree==2.23.4
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
schedule==1.2.2
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
sympy==1.12.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.1
tzdata==2024.1
urllib3==1.25.11
wcwidth==0.2.13
Werkzeug==3.0.4
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.2
